rename <Search/> to <SearchInput/>
