+++ src/components/Home.test.tsx
@@ -5,14 +5,14 @@ jest.mock('react-router-dom/Link');
 jest.mock('@material-ui/core/Button', () => 'button-material-ui');
 jest.mock('@material-ui/core/Grid', () => 'Grid');
 jest.mock('./Icons');
-jest.mock('./Search', () => 'Search');
+jest.mock('./SearchInput', () => 'SearchInput');
 jest.mock('./HomeButtons', () => 'HomeButtons');
 
 describe('<Home/>', () => {
   const wrapper = shallow(<Home />);
 
-  it('renders a <Search/>', () => {
-    expect(wrapper.find('Search').length).toBe(1);
+  it('renders a <SearchInput/>', () => {
+    expect(wrapper.find('SearchInput').length).toBe(1);
   });
 
   it('renders a <HomeButtons/>', () => {
